requirements.txt
fastapi
uvicorn[standard]
qdrant-client
pytest
requests
types-requests
python-dotenv
ruff
black
isort
mypy
